+++ lib/rb_sdl2/rw_ops/rw_file.rb
@@ -1,0 +1,19 @@
+module RbSDL2
+  require_relative 'rw_ops_pointer'
+
+  class RWFile < RWOps
+    def initialize(path, mode = "r")
+      @path = path
+      ptr = ::SDL.RWFromFile(path.to_s, mode.to_s)
+      raise RbSDL2Error if ptr.null?
+      super(RWOpsPointer.new(ptr))
+    end
+
+    def inspect
+      "#<#{self.class.name}:#{path}#{closed? ? " (closed)" : nil}>"
+    end
+
+    attr_reader :path
+    alias to_path path
+  end
+end